requirements.txt
python-telegram-bot==20.6
ccxt>=4.4.82
flask>=3.1.1
nest-asyncio>=1.6.0
pandas>=2.2.3
requests>=2.32.3
schedule>=1.2.2
waitress>=3.0.2
